Refactor: Improve seeding and remove unused routes
The driver routes in `driver.route.ts` were removed as they are no
longer needed. The `IRide` interface in `ride.interface.ts` was updated
to make `_id` optional. Seeding logic was refactored and moved to
dedicated files in the `seed` directory. The `seedSuperAdmin` utility
was removed. The `server.ts` file was updated to use the new seeding
files and run them in parallel.